+++ lib/secret-service/src/dao/secret.spec.js
@@ -225,17 +225,16 @@ describe('SecretDAO', () => {
             value: {
                 authClientId: '5c0a4f796731613b7d10e73e',
                 accessToken,
-                inputFields: JSON.stringify(input),
+                inputFields: input,
             },
             encryptedFields: [],
         }, key, ENCRYPT);
 
-        expect(secret.encryptedFields).toContain('accessToken');
-        expect(secret.encryptedFields).toHaveLength(1);
+        expect(secret.encryptedFields).toEqual(expect.arrayContaining(['accessToken', 'inputFields']));
+        expect(secret.encryptedFields).toHaveLength(2);
 
         expect(secret.value.accessToken).not.toEqual(accessToken);
-        expect(secret.value.inputFields).not.toEqual(input);
-        expect(() => JSON.parse(secret.value.inputFields)).toThrow(Error);
+        expect(secret.value.inputFields).not.toEqual(expect.any(Object));
 
         // full decryption
 
@@ -245,7 +244,7 @@ describe('SecretDAO', () => {
         expect(secret.encryptedFields).toHaveLength(0);
 
         expect(secret.value.accessToken).toEqual(accessToken);
-        expect(JSON.parse(secret.value.inputFields)).toEqual(input);
+        expect(secret.value.inputFields).toMatchObject(input);
 
         // encrypt one field
         secret = SecretDAO.cryptoSecret(secret, key, ENCRYPT, ['accessToken']);
@@ -254,7 +253,7 @@ describe('SecretDAO', () => {
         expect(secret.encryptedFields).toHaveLength(1);
 
         expect(secret.value.accessToken).not.toEqual(accessToken);
-        expect(JSON.parse(secret.value.inputFields)).toEqual(input);
+        expect(secret.value.inputFields).toMatchObject(input);
 
         // decrypt one field
         secret = SecretDAO.cryptoSecret(secret, key, DECRYPT, ['accessToken']);
@@ -263,7 +262,7 @@ describe('SecretDAO', () => {
         expect(secret.encryptedFields).toHaveLength(0);
 
         expect(secret.value.accessToken).toEqual(accessToken);
-        expect(JSON.parse(secret.value.inputFields)).toEqual(input);
+        expect(secret.value.inputFields).toMatchObject(input);
 
         // try to encrypt non existing field
         secret = SecretDAO.cryptoSecret(secret, key, ENCRYPT, ['fooo']);
