+++ lib/secret-service/src/route/secrets/index.spec.js
@@ -13,7 +13,7 @@ const conf = require('../../conf');
 const { AUTH_TYPE, ROLE } = require('../../constant');
 const Server = require('../../server');
 const {
-    SIMPLE, API_KEY, OA2_AUTHORIZATION_CODE, MIXED, SESSION_AUTH,
+    SIMPLE, API_KEY, OA2_AUTHORIZATION_CODE, MIXED,
 } = require('../../constant').AUTH_TYPE;
 const { maskString } = require('../../util/common');
 const dummyUsers = require('../../test/tokens');
@@ -97,12 +97,12 @@ describe('secrets', () => {
         await request.post('/secrets')
             .set(...global.userAuth1)
             .send({
-                    name: 'string',
-                    type: SIMPLE,
-                    value: {
-                        username: 'foo',
-                        passphrase: 'bar',
-                    },
+                name: 'string',
+                type: SIMPLE,
+                value: {
+                    username: 'foo',
+                    passphrase: 'bar',
+                },
             })
             .expect(200);
 
@@ -806,96 +806,6 @@ describe('secrets', () => {
 
         done();
     }, 1000000);
-
-    test('Session Auth setup and request', async () => {
-        const example = nock('https://example.com');
-
-        // nock setup
-        example
-            .persist()
-            .post('/auth')
-            .reply(200, {
-                token: 'letmein',
-            });
-
-        example
-            .get('/userinfo')
-            .reply(200, {
-                sub: 'me',
-            });
-
-        // create auth client
-        const authClient = (await request.post('/auth-clients')
-            .set(...global.userAuth1)
-            .send({
-                type: SESSION_AUTH,
-                name: 'session',
-                fields: [
-                    {
-                        key: 'username',
-                        label: 'Username',
-                        required: true,
-                    },
-                    {
-                        key: 'pass',
-                        label: 'Password',
-                        required: true,
-                    },
-                    {
-                        key: 'subdomain',
-                        label: 'Subdomain',
-                    },
-                ],
-                tokenPath: 'token',
-                endpoints: {
-                    auth: {
-                        label: 'Auth Endpoint',
-                        authType: 'HEADER_AUTH',
-                        url: 'https://example.com/auth',
-                        requestFields: [
-                            {
-                                key: 'user',
-                                value: '{{fields.username}}',
-                            },
-                            {
-                                key: 'password',
-                                value: '{{fields.pass}}',
-                            },
-                        ],
-                    },
-                },
-            })
-            .expect(200)).body.data;
-
-        // Create Session Secret
-        const { _id } = (await request.post('/secrets')
-            .set(...global.userAuth1)
-            .send({
-                name: 'example session secret',
-                owners: [
-                    {
-                        id: '5bc75b5ff231301b6642c854',
-                        type: 'USER',
-                    },
-                ],
-                type: SESSION_AUTH,
-                value: {
-                    authClientId: authClient._id,
-                    inputFields: {
-                        username: 'user1',
-                        pass: 'password1',
-                    },
-                },
-            })
-            .expect(200)).body.data;
-        console.log("new secret: ",_id);
-        // fetch access-token
-        const { body } = await request.get(`/secrets/${_id}`)
-            .set(...global.userToken1ExtraPerm) // userauth1
-            .expect(200);
-
-        expect(body.data.value.accessToken).toEqual('letmein');
-    });
 
     test('Delete secrets with service account', async () => {
         let meta = (await request.get('/secrets/')
